backend1.1/fastapi/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.22
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose==3.5.0
passlib[bcrypt]
python-multipart==0.0.20
